Agrego carga colaboradores y ofrecerProductos

+++ Entrega_3/src/main/java/org/example/Dominio/Colaboraciones/Factory/OfrecerProductosFactory.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 
 public class OfrecerProductosFactory extends ColaboracionFactory {
 
-    public Colaboracion crearColaboracion(Colaborador colaborador, String tipo_producto, String marca, float monto) {
+    public Colaboracion crearColaboracion(Colaborador colaborador, String tipo_producto, String marca, int monto) {
         if(!this.validarTipoColaborador(colaborador, TipoColaborador.P_JURIDICA)){
             return null;
         }
